Class Certificate replaced by class Certificates
Name Certificate was poorly chosen. Better is Certificates so it was
replaced.

Functions were also replaced so that they search every certificate store
in computer if not told otherwise.

In previous version you were obligated to specify specific store, which
you can still do, but if you dont, than every store is searched.